requirements.txt
certifi
charset-normalizer
idna
mysql-connector-python
numpy
pillow
PyQt5
PyQt5-Qt5
PyQt5_sip
requests
staticmap
urllib3
websockets
